requirements.txt
flask
python-dotenv
langchain
sentence-transformers
faiss-cpu
transformers
